Move acceleration controller to common namespace
